--- Backend/Backend.iml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id=":Backend" external.linked.project.path="$MODULE_DIR$" external.root.project.path="$MODULE_DIR$/../forge-1.12.1-14.22.1.2478-mdk" external.system.id="GRADLE" external.system.module.group="forge-1.12.1-14.22.1.2478-mdk" external.system.module.version="unspecified" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/.gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/build" />
-      <excludeFolder url="file://$MODULE_DIR$/gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/out" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="lombok" level="application" />
-  </component>
-</module>
