--- Backend/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Oct 21 12:31:16 BST 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.2.1-all.zip
