+++ src/net/nickac/lithium/backend/controls/impl/LImage.java
