+++ src/net/nickac/lithium/backend/controls/impl/LRadioButton.java
@@ -20,13 +20,14 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
+ *
  */
 
 package net.nickac.lithium.backend.controls.impl;
 
 import net.nickac.lithium.backend.controls.IToggleable;
 import net.nickac.lithium.backend.controls.LControl;
-import net.nickac.lithium.backend.controls.impl.events.AbstractPropertyChangedHandler;
+import net.nickac.lithium.backend.controls.impl.events.PropertyChangedHandler;
 import net.nickac.lithium.backend.other.objects.Color;
 
 import java.util.ArrayList;
@@ -36,8 +37,9 @@ import java.util.UUID;
 /**
  * Created by NickAc for Lithium!
  */
-public class LCheckBox extends LControl implements IToggleable {
-	public LCheckBox(String text) {
+public class LRadioButton extends LControl implements IToggleable {
+
+	public LRadioButton(String text) {
 		setText(text);
 	}
 
@@ -46,7 +48,7 @@ public class LCheckBox extends LControl implements IToggleable {
 	private Color insideColor = Color.BLACK;
 	private Color checkedColor = Color.GRAY;
 	private boolean checked;
-	private transient List<AbstractPropertyChangedHandler<LCheckBox>> toggleHandlers = new ArrayList<>();
+	private transient List<PropertyChangedHandler<LRadioButton>> toggleHandlers = new ArrayList<>();
 
 	@Override
 	public boolean canReceiveUserInput() {
@@ -96,7 +98,7 @@ public class LCheckBox extends LControl implements IToggleable {
 		refresh();
 	}
 
-	public LCheckBox onToggled(AbstractPropertyChangedHandler<LCheckBox> hl) {
+	public LRadioButton onToggled(PropertyChangedHandler<LRadioButton> hl) {
 		toggleHandlers.add(hl);
 		return this;
 	}
