+++ src/net/nickac/lithium/backend/controls/impl/LTextLabel.java
