+++ src/net/nickac/lithium/backend/other/objects/Color.java
