+++ src/net/nickac/lithium/backend/other/objects/Dimension.java
