+++ src/net/nickac/lithium/backend/other/objects/Point.java
