+++ src/net/nickac/lithium/backend/other/objects/Rectangle.java
