+++ src/net/nickac/lithium/backend/other/serverhandlers/LithiumCloseWindowHandler.java
