+++ src/net/nickac/lithium/backend/serializer/SerializationUtils.java
